Made CustomDrawer component
